requirements.txt
biopython>=1.81
anthropic>=0.40.0
python-dotenv>=1.0.0
selenium>=4.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-generativeai>=0.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
gspread>=5.10.0
oauth2client>=4.1.3
certifi>=2024.0.0
